+++ tests/test_date.py
@@ -3,6 +3,7 @@ import datetime
 import pytest
 
 import datacraft
+from datacraft import SpecException
 from . import builder
 
 
@@ -93,7 +94,7 @@ def test_date_center_date():
 
 
 def test_date_center_iso_date():
-    config = {"center_date": "2050-02-15T12:00:00", "stddev_days": 5}
+    config = {"center_date": "2050-02-15T12:00:00Z", "stddev_days": 5}
     spec = _date_iso_spec(**config)
     loader = datacraft.loader.field_loader(spec)
     supplier = loader.get('foo')
@@ -175,6 +176,13 @@ def test_date_restrict_hours():
         date = next(gen)['foo']
         dt = datetime.datetime.strptime(date, date_format)
         assert 6 <= dt.hour <= 21
+
+
+def test_data_and_format_cannot_both_be_specified():
+    date_format = "%d-%m-%Y %H"
+    spec = builder.date(date_format, format=date_format)
+    with pytest.raises(SpecException):
+        datacraft.entries(spec, 1)
 
 
 def _get_unique_values(spec, key, iterations=100):